opcua/requirements.txt
azure-iot-device~=2.7.0
opcua~=0.98.9
python-dotenv~=0.17.1